+++ Babylon Demo/Babylon Demo/ContentView.swift
@@ -2,6 +2,7 @@ import SwiftUI
 
 struct ContentView: View {
     @State var elements: [ListView.Element]?
+    @State private var favourites: [ListView.Element.ID] = [2, 4] // TODO remove harcoded
     
     var body: some View {
         Group {
@@ -21,19 +22,35 @@ extension ContentView {
         url: URL(string: "http://jsonplaceholder.typicode.com/photos")!) { (result: Result<[Photo], RemoteError>) in
             switch result {
             case let .success(photos):
-                self.elements = photos.map {
-                    ListView.Element(
-                        id: $0.id,
-                        title: $0.title,
-                        thumbnail: .init(
-                            image: UIImage(named: "thumbnail_mock")!, // TODO point of sync
-                            size: CGSize(width: 100, height: 100)
-                        )
-                    )
-                }
+                self.elements = photos
+                    .map(self.toElement)
+                    .sorted(by: self.isFavourited)
             case let .failure(error): // TODO use error
                 self.elements = nil
             }
+        }
+    }
+    
+    private func toElement(photo: Photo) ->  ListView.Element{
+        .init(
+            id: photo.id,
+            title: photo.title,
+            thumbnail: .init(
+                image: UIImage(named: "thumbnail_mock")!, // TODO point of sync
+                size: CGSize(width: 100, height: 100)
+            ),
+            isFavourite: self.favourites.contains(where: { photo.id == $0 })
+        )
+    }
+    
+    private func isFavourited(firstElement: ListView.Element, secondElement: ListView.Element) -> Bool {
+        switch (firstElement.isFavourite, secondElement.isFavourite) {
+        case (true, true), (false, false):
+            return firstElement.id < secondElement.id
+        case (true, false):
+            return true
+        case (false, true):
+            return false
         }
     }
 }
@@ -41,21 +58,5 @@ extension ContentView {
 struct ContentView_Previews: PreviewProvider {
     static var previews: some View {
         ContentView(elements: [.fixture(), .fixture(), .fixture()])
-    }
-}
-
-struct Photo {
-    let id: Int
-    let title: String
-    let url: URL
-    let thumbnailURL: URL
-}
-
-extension Photo: Decodable {
-    enum CodingKeys: String, CodingKey {
-        case id
-        case title
-        case url
-        case thumbnailURL = "thumbnailUrl"
     }
 }
