+++ Babylon Demo/Babylon Demo/Preview Content/Preview Assets.xcassets/thumbnail_mock.imageset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "images" : [
+    {
+      "idiom" : "universal",
+      "filename" : "thumbnail_mock.png"
+    }
+  ],
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  }
+}
